requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Gradient Boosting
xgboost==1.7.6
lightgbm==4.0.0

# NLP & Transformers
transformers==4.31.0
sentence-transformers==2.2.2
torch==2.0.1
tokenizers==0.13.3

# Data Processing
nltk==3.8.1
spacy==3.6.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Jupyter
jupyter==1.0.0
notebook==7.0.0
ipywidgets==8.0.7

# Utilities
tqdm==4.65.0
pyyaml==6.0.1
joblib==1.3.1

# Experimentation
statsmodels==0.14.0

# Optional (if downloading datasets programmatically)
kaggle==1.5.16
opendatasets==0.1.22
